Finish minimax algorithm logic. Create new file for board and square components.
